usePostsDataをcustom hooksで切り出し
